requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.51.0
django-environ==0.9.0
django-photologue==3.15.1
django-sortedm2m==3.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
ExifRead==3.0.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.16
